refactor(sdk-provider-solana): address review nits
